+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Support public URL generation
+- Support checksum providing
+- Support temporary URL generation
+
+### Changed
+
+- Move implementation to internal trait `Ajgl\Flysystem\Replicate\ReplicateFilesystemAdapterTrait`
+
 ## 2.1.0 - 2022-10-20
 
 ### Added